requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
Werkzeug
python-dotenv
psycopg2
pandas
